% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biNormCheck.R
\name{biNormCheck}
\alias{biNormCheck}
\title{Bivariate Normality Check}
\usage{
biNormCheck(data, prob)
}
\arguments{
\item{data}{2D input matrix or data frame containing bivariate numerical data}

\item{prob}{The probability used in determining the chi-sq value for degrees of freedom = 2.}
}
\value{
named list with plot of data and the actual vs predicted percentages of data within ellipse
}
\description{
This function evaluates a bivariate data set for normality by observing the predicted and actual probabilities of
containment within a predetermined contour of constant density.
}
\examples{
vec1 <- rnorm(100)
vec2 <- rnorm(100)
data <- data.frame(x=vec1, y=vec2)
biNormCheck(data, prob=0.5)
}
