% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxiestCoxiest.R
\name{boxiestCoxiest}
\alias{boxiestCoxiest}
\title{Boxiest Coxiest}
\usage{
boxiestCoxiest(data)
}
\arguments{
\item{data}{the input data matrix or data frame containing univariate or multivariate data. All input data must be positive}
}
\value{
the augmented data following Box Cox transformation, and a figure showing the respective lambda and likelihood function relations
}
\description{
This function performs the Box Cox transformation on a data frame or data matrix and returns the augmented data.
}
\examples{
vec1 <- rexp(100, rate=2.0)
vec2 <- rexp(100, rate=0.5)
vec3 <- rexp(100, rate=0.1)
data <- data.frame(V1 = vec1, V2 = vec2, V3=vec3)
boxiestCoxiest(data)
}
