% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Detecting Outliers}
\usage{
detectOutliers(data)
}
\arguments{
\item{data}{2D or greater input matrix or data frame containing multivariate numerical data}
}
\value{
a dotplot for each varibles, a scatter plot for each pair of variables, standardized values, and a chi-square plot
}
\description{
This function checks a multivariate numerical data set for outliers by utilizing dot plots, scatter plots,
    standardized values, and chi-square plots.
}
\examples{
vec1 <- rnorm(50)
vec2 <- rnorm(50)
vec3 <- rnorm(50)
vec4 <- rnorm(50)
vec5 <- rnorm(50)
vec6 <- rnorm(50)
vec7 <- rnorm(50)
vec8 <- rnorm(50)
vec9 <- rnorm(50)
data <- data.frame(V1=vec1, V2=vec2, V3=vec3, V4=vec4, V5=vec5, V6=vec6, V7=vec7, V8=vec8, V9=vec9)
detectOutliers(data)
}
