% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvNormCheck.R
\name{mvNormCheck}
\alias{mvNormCheck}
\title{Multivariate Normality Check}
\usage{
mvNormCheck(data)
}
\arguments{
\item{data}{2D or greater input matrix or data frame containing multivariate numerical data}
}
\value{
a chi-square plot
}
\description{
This function checks a multivariate numerical data set for normality by utilizing the squared generalized distances
and their relationship to specific chi-square values.
}
\examples{
vec1 <- rnorm(100)
vec2 <- rnorm(100)
vec3 <- rnorm(100)
data <- data.frame(x=vec1, y=vec2, z=vec3)
mvNormCheck(data)
}
