% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{QQplot}
\alias{QQplot}
\title{QQ Plot}
\usage{
QQplot(data)
}
\arguments{
\item{data}{1D input vector containing numerical data}
}
\value{
a QQ plot of the input data
}
\description{
This function creates a QQ plot of the input data.
}
\examples{
data <- rnorm(1000)
QQplot(data)
}
