% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQ.R
\name{rQ}
\alias{rQ}
\title{Correlation Coefficient (rQ)}
\usage{
rQ(data)
}
\arguments{
\item{data}{1D input vector containing numerical data}
}
\value{
the rQ value (correlation coefficient) of the ordered data and standard normal quantiles
}
\description{
This function determines the rQ value for a dataset to aid in the determination of normality.
}
\examples{
data <- rnorm(1000)
QQplot(data)
}
