% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniNormCheck.R
\name{uniNormCheck}
\alias{uniNormCheck}
\title{Check Univariate Normality}
\usage{
uniNormCheck(data)
}
\arguments{
\item{data}{1D input vector or list containing data}
}
\value{
True if data is normal and False if data is not normal
}
\description{
This function will check a 1 dimensional dataset for normality. Please note that this function utilizes
an empiral test and is not as accurate as functions like shapiro.test() which provides a more formal assessment of
the data.
}
\examples{
data <- rnorm(1000)
uniNormCheck(data)
}
